chore: upgrade to TS 4.7, compile with NodeNext
