+++ __tests__/validate-tsconfig.test.ts
@@ -40,14 +40,12 @@ const tsconfigSchema = Joi.object({
         incremental: Joi.forbidden(),
         tsBuildInfoFile: Joi.forbidden(),
         outDir: Joi.forbidden(),
-        module: Joi.valid('commonjs', 'es2020', 'esnext').required(),
       }).unknown(),
       otherwise: Joi.object({
         noEmit: Joi.valid(false).required(),
         incremental: Joi.valid(true).required(),
         rootDir: Joi.valid('src').required(),
         outDir: Joi.valid('lib').required(),
-        module: Joi.valid('commonjs', 'es2020', 'esnext').required(),
       }).unknown(),
     },
   ),
