+++ packages/create-docusaurus/templates/classic-typescript/package.json
@@ -26,7 +26,7 @@
   "devDependencies": {
     "@docusaurus/module-type-aliases": "2.0.0-beta.21",
     "@tsconfig/docusaurus": "^1.0.5",
-    "typescript": "^4.6.4"
+    "typescript": "^4.7.3"
   },
   "browserslist": {
     "production": [
