+++ packages/create-docusaurus/tsconfig.json
@@ -3,7 +3,6 @@
   "references": [{"path": "./tsconfig.build.json"}],
   "compilerOptions": {
     "noEmit": true,
-    "module": "esnext",
     "rootDir": "."
   },
   "include": ["bin"],
