+++ packages/docusaurus-logger/tsconfig.json
@@ -6,7 +6,6 @@
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
     "sourceMap": true,
     "declarationMap": true,
-    "module": "commonjs",
     "rootDir": "src",
     "outDir": "lib"
   },
