+++ packages/docusaurus-migrate/tsconfig.build.json
@@ -5,7 +5,6 @@
     "composite": true,
     "incremental": true,
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
-    "module": "commonjs",
     "rootDir": "src",
     "outDir": "lib"
   },
