+++ packages/docusaurus-plugin-client-redirects/src/deps.d.ts
@@ -5,10 +5,6 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-declare module '@docsearch/react/modal';
-
-declare module '@docsearch/react/style';
-
 // TODO incompatible declaration file
 declare module 'eta' {
   export const defaultConfig: object;
