+++ packages/docusaurus-plugin-content-docs/tsconfig.json
@@ -5,7 +5,6 @@
     "noEmit": false,
     "incremental": true,
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
-    "module": "commonjs",
     "rootDir": "src",
     "outDir": "lib"
   },
