+++ packages/docusaurus-plugin-content-pages/tsconfig.json
@@ -4,7 +4,6 @@
     "noEmit": false,
     "incremental": true,
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
-    "module": "commonjs",
     "rootDir": "src",
     "outDir": "lib"
   },
