+++ packages/docusaurus-theme-translations/tsconfig.build.json
@@ -7,7 +7,6 @@
     "tsBuildInfoFile": "./lib/.tsbuildinfo",
     "sourceMap": true,
     "declarationMap": true,
-    "module": "commonjs",
     "rootDir": "src",
     "outDir": "lib"
   },
