+++ packages/docusaurus-theme-translations/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../../tsconfig.json",
   "references": [{"path": "./tsconfig.build.json"}],
   "compilerOptions": {
-    "module": "esnext",
     "noEmit": true,
     "checkJs": true
   },
